Update network parameter format to support network IDs in MAC addresses
This commit enhances the --networks parameter in spawn-vm.sh to accept networkname:id format, where the ID is used as part of the MAC address generation.

Key changes:
- Modified --networks parameter to accept format 'networkname:id'
- Updated MAC address generation to use network ID as the last byte
- MAC format is now: :::
- Improved help message with clear examples of the new parameter format
- Added fallback to use index as ID if none is provided

This change makes it easier to identify VMs when enlisting them in MAAS by creating predictable MAC addresses based on network IDs.